backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.14.0,<2.0.0
pydantic[email]>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
aiosmtplib>=3.0.0,<4.0.0
jinja2>=3.1.0,<4.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0,<4.0.0
